add generic message formatting for bot
