+++ discovery-provider/plugins/verified-uploads/src/handlers/users.js
@@ -1,0 +1,10 @@
+export default async ({ slack, db }, { user_id }) => {
+  const user = await db("users")
+    .select("handle", "name")
+    .where("user_id", "=", user_id)
+    .first()
+    .catch(console.error);
+
+  if (user) {
+  }
+};
