+++ discovery-provider/plugins/verified-uploads/src/slack.js
@@ -4,10 +4,24 @@ export default () => {
   const { SLACK_TOKEN, SLACK_CHANNEL } = process.env;
   const web = new WebClient(SLACK_TOKEN);
   return {
-    sendMsg: (msg) =>
-      web.chat.postMessage({
+    sendMsg: (header, body) => {
+      const msg = `${header} ${formatter(body)}`;
+      return web.chat.postMessage({
         text: msg,
         channel: SLACK_CHANNEL,
-      }),
+      });
+    },
   };
+};
+
+const formatter = (data) => {
+  const msg = [];
+  for (const [key, value] of Object.entries(data)) {
+    // omit any null entries of the track
+    if (value != null) {
+      msg.push(`${key}: ${value}`);
+    }
+  }
+  const inner = msg.join("\n");
+  return "```" + inner + "```";
 };
